feat: added deps + added x-on-resize
With this change, a new daemon is added. This daemon listens for resize
events on the root display area. Upon a resize event, it calls `xrandr
--auto`.

There's the possibility to call some external program. My idea is to
make use of:
- WiFi SSID to which machine is connected
- extenal display port that shows as connected

In this way, I can write a script that will issue the right xrandr
command depending on where I am and which port is connected.